lfs: add a list of symlinks to all needed Lua files
It's important that all files remain in the same directory and are
built with no directory component in them otherwise it will appear
into the output file. Some files come from the project itself and
others from the iot-core one supposed to be placed in "..".

A link to lfs.img is made to ease uploading (which also needs not to
have any component directory) without having to "cd .." or "cd -"
between each and every command.